BRB register, team list, directions schedule pages
Auto redirect shsrobotics.us/brb.html to shsrobotics.us/brb since the
latter was on a Chief Delphi post.
Put a copy of brb-flyer.jpg file in the original /img/ folder instead
of just in /assets/img/ because it
was linked on a Chief Delphi post. TODO: delete this copy after BRB 2019