Make BRB registration form less confusing
I noticed that with the current size of the BRB form's visible area, it
may not be clear that there is more to registering than just sending in
the registration fee check (since only the registration fee instructions
are visible, and the cutoff is right near the end of the instructions).
There is a concern that some teams may have missed this and only mailed
checks, though this needs to be confirmed (I do not know yet if any
registration fees have been received). I also reworded the instructions
on the google form itself that the form needs to be filled out in
addition to paying the registration fee.